Fix error in award_star when more than one star awarded.
Fixed error in award_star method when more than one star awarded the
remaining points were being saved on current smiley in iteration causing
those points to be lost for further calculation during the loop.
Now if we are at the last item in the queryset and there are any
remaining points they get saved properly on the last item.
Added tests confirming that.