Make fonts local.
Found issue with hardcoded paths to font files when uploaded to staging.
The files are not properly resolved by amazon s3 as no static call is
made.
Therefore I created load_fonts.html include where all the fonts are
declared and need to be loaded in the <head> of the base template.
Moved all the fonts to static/core/fonts/
Added load_fonts.html as an include in base.html.